use utils to fetch data in the CheckCompanyIdentifierView endpoint
